add horizontal movement

+++ src/physics.rs
@@ -7,10 +7,14 @@ impl Plugin for PhysicsPlugin {
         app.add_systems(
             (ground_detection, apply_gravity, apply_velocity)
                 .chain()
+                .in_set(PhysicsSet)
                 .in_schedule(CoreSchedule::FixedUpdate),
         );
     }
 }
+
+#[derive(SystemSet, Debug, Hash, PartialEq, Eq, Clone)]
+pub struct PhysicsSet;
 
 /// Direction gravity applies to for a specific object,
 /// Note: might be better for this to be a vector instead?
@@ -27,13 +31,22 @@ pub enum GravityDirection {
 pub struct Ground;
 
 impl GravityDirection {
-    fn as_vec2(&self) -> Vec2 {
+    pub fn as_vec2(&self) -> Vec2 {
         match self {
             GravityDirection::Down => Vec2::NEG_Y,
             GravityDirection::Up => Vec2::Y,
             GravityDirection::Left => Vec2::NEG_X,
             GravityDirection::Right => Vec2::X,
         }
+    }
+
+    pub fn forward(&self) -> Vec2 {
+      match self {
+        GravityDirection::Down => Vec2::X,
+        GravityDirection::Up => Vec2::NEG_X,
+        GravityDirection::Left => Vec2::Y,
+        GravityDirection::Right => Vec2::NEG_Y,
+      }
     }
 }
 
@@ -49,7 +62,8 @@ fn apply_gravity(
     mut q: Query<(&mut Velocity, &GravityDirection, &JumpState)>,
     time: Res<FixedTime>,
 ) {
-    const GRAVITY: f32 = 500.0;
+    const GRAVITY_UP: f32 = 500.0;
+    const GRAVITY_DOWN: f32 = 750.0;
     for (mut v, dir, jump_state) in q.iter_mut() {
         if jump_state.on_ground {
             // zero velocity in gravity direction. probably need to add some thresholds here too.
@@ -57,7 +71,12 @@ fn apply_gravity(
             continue;
         }
 
-        v.0 += GRAVITY * time.period.as_secs_f32() * dir.as_vec2();
+        let gravity = if dir.as_vec2().dot(v.0) < 0. {
+            GRAVITY_UP
+        } else {
+            GRAVITY_DOWN
+        };
+        v.0 += gravity * time.period.as_secs_f32() * dir.as_vec2();
     }
 }
 
@@ -69,11 +88,24 @@ fn apply_velocity(mut query: Query<(&mut Transform, &Velocity)>, time_step: Res<
 }
 
 fn ground_detection(
-    mut jumpers: Query<(&mut JumpState, &Transform, &GravityDirection, &Sprite), Without<Ground>>,
+    mut jumpers: Query<
+        (
+            &mut JumpState,
+            &Velocity,
+            &Transform,
+            &GravityDirection,
+            &Sprite,
+        ),
+        Without<Ground>,
+    >,
     grounds: Query<Entity, With<Ground>>,
     rapier: Res<RapierContext>,
 ) {
-    for (mut j, t, g, s) in &mut jumpers {
+    for (mut j, v, t, g, s) in &mut jumpers {
+        // only check ground detection when moving in the same direction as gravity
+        if g.as_vec2().dot(v.0) < 0.0 {
+            continue;
+        }
         let Some(sprite_size) = s.custom_size else { panic!("cannot get size of sprite") };
         // left ray when gravity is down
         let ray_origin_1 = t.translation.truncate() - sprite_size / 2.;
